requirements.txt
django-jsonfield==0.9.13
django-model-utils==2.2
oauth2client==1.3.1
sphinx==1.2.3
stripe==1.19.1
